--- src/js/purchase/disabledPurchaseButton.js
@@ -1,10 +1,0 @@
-import { $ } from '../../util/selector.js';
-
-const disablePurchaseButton = () => {
-  const purchaseButton = $('#purchase-form button');
-  purchaseButton.disabled = true;
-  purchaseButton.style.backgroundColor = '#ccc';
-  purchaseButton.style.cursor = 'not-allowed';
-};
-
-export default disablePurchaseButton;
