+++ src/js/submitWinningNumberForm.js
@@ -5,9 +5,7 @@ import { $, $all } from '../util/selector.js';
 import validateBonusNumber from '../validation/validateBonusNumber.js';
 import validateWinningNumber from '../validation/validateWinningNumber.js';
 import { resetError, showError } from './errorHandler.js';
-import { lockScroll } from './scroll.js';
 
-// ✅ 2. 당첨 번호 입력 폼 이벤트 설정
 export const submitWinningNumberForm = async () => {
   return new Promise((resolve) => {
     $('#winning-number-form').addEventListener('submit', (event) => {
@@ -15,13 +13,10 @@ export const submitWinningNumberForm = async () => {
       const winningLotto = handleWinningNumber();
 
       resolve(winningLotto);
-      // const matchingResult = calculateMatchingResult(winningLotto, lottoArray);
-      // const profitRate = calculateProfitRate(matchingResult, lottoArray.length);
     });
   });
 };
 
-// 🔹 2. 당첨 번호 입력 처리
 const handleWinningNumber = () => {
   const modal = $('#result-modal');
   const errorUI = $('#winning-number-error');
@@ -32,7 +27,6 @@ const handleWinningNumber = () => {
 
     const winningLotto = new WinningLotto(new Lotto(winningNumbers), bonusNumber);
 
-    showModal(modal);
     return winningLotto;
   } catch (error) {
     showError(errorUI, error.message);
@@ -53,10 +47,4 @@ const getWinningNumbers = () => {
   const bonusNumber = parseBonusNumber(bonusNumberInput);
 
   return { winningNumbers, bonusNumber };
-};
-
-// 🔹 3. 모달 관련 함수
-const showModal = (modal) => {
-  modal.style.display = 'flex';
-  lockScroll();
 };
